backend/requirements.txt
Django==6.0.3
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
